+++ packages/manager/src/components/PrimaryNav/PrimaryNav.tsx
@@ -34,6 +34,7 @@ import {
   useObjectStorageBuckets,
   useObjectStorageClusters,
 } from 'src/queries/objectStorage';
+import { useRegionsQuery } from 'src/queries/regions/regions';
 import { useStackScriptsOCA } from 'src/queries/stackscripts';
 import { isFeatureEnabled } from 'src/utilities/accountCapabilities';
 
@@ -98,6 +99,18 @@ export const PrimaryNav = (props: PrimaryNavProps) => {
 
   const { _isManagedAccount, account, accountError } = useAccountManagement();
 
+  const isObjMultiClusterEnabled = isFeatureEnabled(
+    'Object Storage Access Key Regions',
+    Boolean(flags.objMultiCluster),
+    account?.capabilities ?? []
+  );
+
+  const { data: regions } = useRegionsQuery();
+
+  const regionsSupportingObjectStorage = regions?.filter((region) =>
+    region.capabilities.includes('Object Storage')
+  );
+
   const {
     data: oneClickApps,
     error: oneClickAppsError,
@@ -108,13 +121,28 @@ export const PrimaryNav = (props: PrimaryNavProps) => {
     data: clusters,
     error: clustersError,
     isLoading: clustersLoading,
-  } = useObjectStorageClusters(enableObjectPrefetch);
+  } = useObjectStorageClusters(
+    enableObjectPrefetch && !isObjMultiClusterEnabled
+  );
 
+  /*
+   @TODO OBJ Multicluster:'region' will become required, and the
+   'cluster' field will be deprecated once the feature is fully rolled out in production.
+   As part of the process of cleaning up after the 'objMultiCluster' feature flag, we will
+   remove 'cluster' and retain 'regions'.
+  */
   const {
     data: buckets,
     error: bucketsError,
     isLoading: bucketsLoading,
-  } = useObjectStorageBuckets(clusters, enableObjectPrefetch);
+  } = useObjectStorageBuckets({
+    clusters: isObjMultiClusterEnabled ? undefined : clusters,
+    enabled: enableObjectPrefetch,
+    isObjMultiClusterEnabled,
+    regions: isObjMultiClusterEnabled
+      ? regionsSupportingObjectStorage
+      : undefined,
+  });
 
   const allowObjPrefetch =
     !buckets &&
@@ -173,14 +201,6 @@ export const PrimaryNav = (props: PrimaryNavProps) => {
           icon: <Linode />,
         },
         {
-          betaChipClassName: 'beta-chip-placement-groups',
-          display: 'Placement Groups',
-          hide: !flags.placementGroups?.enabled,
-          href: '/placement-groups',
-          icon: <PlacementGroups />,
-          isBeta: flags.placementGroups?.beta,
-        },
-        {
           display: 'Volumes',
           href: '/volumes',
           icon: <Volume />,
@@ -221,6 +241,14 @@ export const PrimaryNav = (props: PrimaryNavProps) => {
           display: 'Images',
           href: '/images',
           icon: <Image />,
+        },
+        {
+          betaChipClassName: 'beta-chip-placement-groups',
+          display: 'Placement Groups',
+          hide: !flags.placementGroups?.enabled,
+          href: '/placement-groups',
+          icon: <PlacementGroups />,
+          isBeta: flags.placementGroups?.beta,
         },
       ],
       [