+++ packages/manager/src/features/PlacementGroups/PlacementGroupsDetail/PlacementGroupsSummary/PlacementGroupsSummary.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { placementGroupFactory } from 'src/factories';
+import { placementGroupFactory, regionFactory } from 'src/factories';
 import { renderWithTheme } from 'src/utilities/testHelpers';
 
 import { PlacementGroupsSummary } from './PlacementGroupsSummary';
@@ -36,8 +36,10 @@ describe('PlacementGroups Summary', () => {
               linode_id: 10,
             },
           ],
-
           region: 'us-east',
+        })}
+        region={regionFactory.build({
+          id: 'us-east',
         })}
       />
     );