+++ packages/manager/src/queries/vlans.ts
@@ -1,18 +1,40 @@
-import { APIError } from '@linode/api-v4/lib/types';
-import { VLAN, getVlans } from '@linode/api-v4/lib/vlans';
-import { useQuery } from '@tanstack/react-query';
+import { getVlans } from '@linode/api-v4/lib/vlans';
+import { createQueryKeys } from '@lukemorales/query-key-factory';
+import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
 
 import { getAll } from 'src/utilities/getAll';
 
-import { queryPresets } from './base';
+import type { APIError, Filter, ResourcePage, VLAN } from '@linode/api-v4';
 
-export const queryKey = 'vlans';
-
-const _getVlans = (): Promise<VLAN[]> =>
+const getAllVLANs = (): Promise<VLAN[]> =>
   getAll<VLAN>((params) => getVlans(params))().then(({ data }) => data);
 
+export const vlanQueries = createQueryKeys('vlans', {
+  all: {
+    queryFn: getAllVLANs,
+    queryKey: null,
+  },
+  infinite: (filter: Filter = {}) => ({
+    queryFn: ({ pageParam = 1 }) =>
+      getVlans({ page: pageParam, page_size: 25 }, filter),
+    queryKey: [filter],
+  }),
+});
+
 export const useVlansQuery = () => {
-  return useQuery<VLAN[], APIError[]>([queryKey], _getVlans, {
-    ...queryPresets.longLived,
+  return useQuery<VLAN[], APIError[]>(vlanQueries.all);
+};
+
+export const useVLANsInfiniteQuery = (filter: Filter = {}, enabled = true) => {
+  return useInfiniteQuery<ResourcePage<VLAN>, APIError[]>({
+    getNextPageParam: ({ page, pages }) => {
+      if (page === pages) {
+        return undefined;
+      }
+      return page + 1;
+    },
+    keepPreviousData: true,
+    ...vlanQueries.infinite(filter),
+    enabled,
   });
 };