+++ packages/manager/src/utilities/pricing/constants.ts
@@ -6,7 +6,6 @@ export interface ObjStoragePriceObject {
 
 // These values will eventually come from the API, but for now they are hardcoded and
 // used to generate the region based dynamic pricing.
-export const NODEBALANCER_PRICE = 10;
 export const LKE_HA_PRICE = 60;
 export const OBJ_STORAGE_PRICE: ObjStoragePriceObject = {
   monthly: 5.0,